# SWAGGER-NVSL-APIS.yaml
openapi: 3.0.3
info:
  title: NVSL API
  version: 1.0.0
  description: |
    Documentação consolidada das APIs do NVSL.
    Base URL padrão: `http://localhost:8081/api`.

    Rotas **web** na raiz da aplicação (`http://localhost:8081`, cadastro MGI):
    - `GET /` — status da API
    - `GET /redirect-gov` — callback OAuth2/OIDC (redirect_uri)
    - `GET /logout` — redireciona ao login do frontend (`?from=logout`); revogação do token: `POST /api/auth/logout`
servers:
  - url: http://localhost:8081/api
    description: Ambiente local (Docker)
...
paths:
...
  /localidades/municipios:
    get:
      tags: [Localidades]
      summary: Lista municípios por UF
      parameters:
        - name: uf
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 2
          example: SP
      responses:
        "200":
          description: Lista de municípios da UF
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Option"
        "422":
          description: UF inválida
...
  /solicitacoes-cadastro:
    get:
      tags: [Solicitações de Cadastro]
      summary: Lista solicitações de cadastro (com filtros)
      security:
        - bearerAuth: []
      parameters:
        - name: cpf
          in: query
          schema: { type: string }
        - name: nome
          in: query
          schema: { type: string }
        - name: uf
          in: query
          schema: { type: string }
        - name: municipio
          in: query
          schema: { type: string }
        - name: orgao
          in: query
          schema: { type: string }
        - name: esfera
          in: query
          schema:
            type: string
            enum: [federal, estadual, municipal]
        - name: status
          in: query
          schema:
            type: string
            enum: [em_analise, aprovado, reprovado]
      responses:
        "200":
          description: Lista de solicitações
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SolicitacaoItem"
        "401":
          description: Não autenticado

    post:
      tags: [Solicitações de Cadastro]
      summary: Cria solicitação de cadastro (público)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitacaoCadastroCreateRequest"
      responses:
        "201":
          description: Solicitação criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolicitacaoCadastroCreateResponse"
        "422":
          description: Erro de validação
  /solicitacoes-cadastro/verificar-cpf:
    get:
      tags: [Solicitações de Cadastro]
      summary: Verifica disponibilidade de CPF
      parameters:
        - name: cpf
          in: query
          required: true
          schema:
            type: string
          example: "12345678909"
      responses:
        "200":
          description: Resultado da validação do CPF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificarCpfResponse"
        "422":
          description: CPF inválido
  /solicitacoes-cadastro/{id}:
    get:
      tags: [Solicitações de Cadastro]
      summary: Detalha uma solicitação de cadastro
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Detalhes da solicitação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolicitacaoDetalhe"
        "401":
          description: Não autenticado
        "403":
          description: Acesso negado
        "404":
          description: Solicitação não encontrada

    patch:
      tags: [Solicitações de Cadastro]
      summary: Aprova/Reprova solicitação
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SolicitacaoUpdateRequest"
      responses:
        "200":
          description: Solicitação atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SolicitacaoUpdateResponse"
        "401":
          description: Não autenticado
        "403":
          description: Acesso negado
        "404":
          description: Solicitação não encontrada
        "422":
          description: Erro de validação/regra de negócio
  /solicitacoes-cadastro/{id}/perfis/{perfilUsuarioId}/ativar:
    patch:
      tags: [Solicitações de Cadastro]
      summary: Ativa cadastro/perfil vinculado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: perfilUsuarioId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Perfil ativado
        "401":
          description: Não autenticado
        "403":
          description: Acesso negado
        "404":
          description: Recurso não encontrado
  /solicitacoes-cadastro/{id}/perfis:
    post:
      tags: [Solicitações de Cadastro]
      summary: Adiciona novo perfil vinculado à solicitação
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdicionarPerfilVinculadoRequest"
      responses:
        "201":
          description: Perfil vinculado adicionado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  data:
                    type: object
        "401":
          description: Não autenticado
        "403":
          description: Acesso negado
        "404":
          description: Solicitação não encontrada
        "422":
          description: Erro de validação
  /solicitacoes-cadastro/{id}/perfis/{perfilUsuarioId}/desativar:
    patch:
      tags: [Solicitações de Cadastro]
      summary: Desativa cadastro/perfil vinculado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: perfilUsuarioId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Perfil desativado
        "401":
          description: Não autenticado
        "403":
          description: Acesso negado
        "404":
          description: Recurso não encontrado
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
    Option:
      type: object
      properties:
        value: { type: string }
        label: { type: string }
...
    VerificarCpfResponse:
      type: object
      properties:
        disponivel: { type: boolean }
        mensagem: { type: string }
    SolicitacaoCadastroCreateRequest:
      type: object
      required:
        - nome
        - CPF
        - emailInstitucional
        - telefoneInstitucional
        - esferaAtuacao
        - uf
        - municipio
        - orgao
        - cargo
      properties:
        nome: { type: string, example: "Fulano da Silva" }
        CPF: { type: string, example: "12345678909" }
        emailInstitucional: { type: string, example: "fulano@email.com" }
        telefoneInstitucional: { type: string, example: "6133334444" }
        telefonePessoal: { type: string, nullable: true, example: "61999998888" }
        esferaAtuacao:
          type: string
          enum: [federal, estadual, municipal]
        uf: { type: string, example: "DF" }
        municipio: { type: string, example: "Brasília" }
        orgao: { type: string, example: "Ministério dos Direitos Humanos" }
        cargo: { type: string, example: "Analista" }
        aceiteTermo:
          type: boolean
          description: Opcional. A ciência do termo é implícita no envio; o backend registra o horário em aceite_termo_at.
    SolicitacaoCadastroCreateResponse:
      type: object
      properties:
        message: { type: string }
        solicitacao_id: { type: integer }
    SolicitacaoItem:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time, nullable: true }
        cpf: { type: string }
        esfera_atuacao: { type: string }
        uf: { type: string }
        municipio: { type: string }
        orgao: { type: string }
    SolicitacaoDetalhe:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        cpf: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
        email_institucional: { type: string }
        telefone_institucional: { type: string }
        telefone_pessoal: { type: string, nullable: true }
        esfera_atuacao: { type: string }
        uf: { type: string }
        municipio: { type: string }
        orgao: { type: string }
        cargo: { type: string, nullable: true }
        perfis_vinculados:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              perfil: { type: string }
              vigencia_inicio: { type: string }
              vigencia_fim: { type: string }
              vigente: { type: boolean }
              esfera: { type: string }
              uf: { type: string }
              municipio: { type: string }
              orgao: { type: string }
              cargo: { type: string }
    SolicitacaoUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [aprovado, reprovado]
        perfil_id:
          type: integer
          description: Obrigatório quando status=aprovado
        vigencia_inicio:
          type: string
          format: date
          nullable: true
        vigencia_fim:
          type: string
          format: date
          nullable: true
    SolicitacaoUpdateResponse:
      type: object
      properties:
        message: { type: string }
        data:
          type: object
          properties:
            id: { type: integer }
            status: { type: string }
    AdicionarPerfilVinculadoRequest:
      type: object
      required: [perfil_id, vigencia_inicio]
      properties:
        perfil_id:
          type: integer
          example: 1
        vigencia_inicio:
          type: string
          format: date
          example: "2026-03-25"
        vigencia_fim:
          type: string
          format: date
          nullable: true
          example: "2026-12-31"
        orgao:
          type: string
          nullable: true
        cargo:
          type: string
          nullable: true